Fixed GSKA, it had two important errors.
The random individual was selected for each dimension, it should be the same
for all of them.

The best and worse individual were not changed, now they are.

An example, for previous version, for CEC'2017 with dimensión 10, gived an error
of 13144, with the change it gives an error of 1e-12 (now it is consistent with
the results in the original paper).